Quick implementation of Gps
added all permissions and location gathering features, edited item object to hold LocationResult type. i want to see if i can convert it to a string and add it to the switch list.
currently does not set the item objects value for gps to anything. Created set gps setter.
The GPS permissions request is on the searchPage since that is defaulted to our *home* page. we need to request every application launch unless phone setting are specified to not require it more than once.
Ill do the gps info setting after walk.